move helper to the correct package
