+++ core/src/test/scala/org/dbpedia/spotlight/db/similarity/TestGenerativeContextSimilarity.scala
@@ -2,7 +2,8 @@ package org.dbpedia.spotlight.db.similarity
 
 import org.dbpedia.spotlight.db.memory.MemoryQuantizedCountStore
 import org.dbpedia.spotlight.model.{DBpediaResource, TokenType}
-import org.dbpedia.spotlight.Helper.{createContextStore, createTokenTypeStore}
+import org.dbpedia.spotlight.util.MemoryStoreUtil
+import MemoryStoreUtil.{createContextStore, createTokenTypeStore}
 import org.junit.Assert._
 import org.junit.Test
